% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-base.R
\name{bb_sims_base}
\alias{bb_sims_base}
\title{Simulate Boreal Caribou Data from Base Parameter}
\usage{
bb_sims_base(
  nyear,
  month_composition,
  population_init,
  proportion_adult_female,
  proportion_yearling_female,
  survival_month_year,
  fecundity_year,
  collared_adult_females,
  month_collar,
  probability_uncertain_mortality_month_year,
  probability_uncertain_survival_month_year,
  group_size_lambda_year,
  group_size_theta_year,
  group_coverage_year,
  group_min_size,
  group_max_proportion,
  probability_unsexed_adult_female_year,
  probability_unsexed_adult_male_year
)
}
\arguments{
\item{nyear}{The number of caribou years.}

\item{month_composition}{A number between 1 and 12 of the caribou month that composition survey occurs relative to the calving month.}

\item{population_init}{A vector of length 3 corresponding to the population of female calves, female yearlings and female adults at the start of the first year.}

\item{survival_month_year}{An array of dimensions 12 x nyear x 3. The third dimension represents female calf, female yearling and female adult stages.}

\item{fecundity_year}{A matrix of of dimensions nyear x 3. The second dimension corresponds to female calf, female yearling and female adult stages.}

\item{collared_adult_females}{A whole positive number of the number of collared adult females. Each year the total collared adult females will be topped up to equal this number.}

\item{probability_uncertain_mortality_month_year}{A matrix of dimensions 12 x nyear of the monthly probability of the fate of a mortality being uncertain.}

\item{group_size_lambda_year}{A vector of length nyear of the expected group size.}

\item{group_size_theta_year}{A vector of length nyear of the standard deviation of the extra-Poisson variation in the group size.}

\item{probability_unsexed_adult_female_year}{A vector of length nyear of the probability of an adult female being unsexed.}

\item{probability_unsexed_adult_male_year}{A vector of length nyear of the probability of an adult male being unsexed.}

\item{proportion_adult_females}{A number between 0 and 1 of the proportion of adults that are female.}

\item{proportion_yearling_females}{A number between 0 and 1 of the proportion of adults that are female. This applies to calf sex ratio as well.}

\item{probability_uncertain_survivor_month_year}{A matrix of dimensions 12 x nyear of the monthly probability of the fate of a survivor being uncertain.}

\item{groups_coverage_year}{A vector of length nyear of the proportion of the groups in the composition survey.}
}
\value{
A list of three tibbles.
The first named survival has columns year, month, starttotal, mortalitiescertain and mortalitiesuncertain.
The second named recruitment has columns year, month, cows, bulls, unknownadults, yearlings and calves.
And the final named abundance has columns year, month, sex, stage, abundance.
}
\description{
The survival for a month is the survival from the start to the end of a month.
Except for the starting values, the abundance for a month is the abundance at the end of the month.
The number of collared females of a life stage is the minimum of the
number specified and the number of individuals available.
Individual are assigned to composition groups based on their relative proportions
until there are no individuals remaining.
Cows, bulls and yearlings are assigned to groups based on their relative proportions.
If a cow is assigned and there is space for one more individual then a calf is a
For each cow (or yearling if have calves) probability of having a calf.
The model includes demographic stochasticity.
}
\examples{
bb_sims_base()
}
