% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbousims-package.R
\docType{package}
\name{bbousims-package}
\alias{bbousims}
\alias{bbousims-package}
\title{bbousims: Simulate Boreal Caribou Survival, Recruitment and Population Growth}
\description{
Simulates survival and recruitment data for boreal caribou populations using hierarchical Bayesian models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/bbousims}
  \item \url{https://poissonconsulting.github.io/bbousims/}
  \item Report bugs at \url{https://github.com/poissonconsulting/bbousims/issues}
}

}
\author{
\strong{Maintainer}: Seb Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})
  \item John Boulanger (\href{https://orcid.org/0000-0001-8222-1445}{ORCID})
  \item Ayla Pearson \email{ayla@poissonconsulting.ca}
}

Other contributors:
\itemize{
  \item Environment and Climate Change Canada [copyright holder]
}

}
\keyword{internal}
