% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth.R
\name{bbs_fecundity_caribou}
\alias{bbs_fecundity_caribou}
\title{Get stochastic Boreal Caribou fecundity rates by year and stage.}
\usage{
bbs_fecundity_caribou(
  calves_per_adult_female,
  trend = 0,
  annual_sd = 0,
  nyear = 10
)
}
\arguments{
\item{calves_per_adult_female}{A number of the calves per adult female.}

\item{trend}{A number of the effect of an increase of one year on the log-odds calves per adult female.}

\item{annual_sd}{A number of the standard deviation of the annual variation on the log-odds calves per adult female.}

\item{nyear}{A whole number of the number of years.}
}
\value{
A matrix of fecundity rates with dimensions year and stage.
}
\description{
Year is scaled to Year - 1 for trend, which ensures that \code{calves_per_adult_female} is the rate in the first year.
}
\examples{
if(interactive()){
  bbs_fecundity_caribou(0.4, trend = 0.1, annual_sd = 0.3)
}

}
