% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth-matrix.R
\name{bbs_matrix_birth_year}
\alias{bbs_matrix_birth_year}
\title{Create a birth process matrix for each year.}
\usage{
bbs_matrix_birth_year(
  fecundity,
  female_recruit_stage = 1,
  male_recruit_stage = NULL,
  proportion_female = 0.5
)
}
\arguments{
\item{fecundity}{A matrix of the fecundity rates with dimensions year and stage.}

\item{female_recruit_stage}{A positive whole number of the stage representing female recruits.}

\item{male_recruit_stage}{A positive whole number of the stage representing male recruits. Ignored if NULL.}

\item{proportion_female}{A number between 0 and 1 indicating the proportion of recruits that are female.}
}
\value{
An array of the birth process matrices with dimensions stage, stage, year.
}
\description{
Create a birth process matrix for each year.
}
\examples{
fec <- bbs_fecundity(c(NA, 0.7), nyear = 3)
bbs_matrix_birth_year(fec$eFecundity)
}
