% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-matrix.R
\name{bbs_matrix_survival}
\alias{bbs_matrix_survival}
\title{Create a survival process matrix.}
\usage{
bbs_matrix_survival(survival)
}
\arguments{
\item{survival}{A vector of the survival rates in each stage.}
}
\value{
A matrix of the survival subprocess.
}
\description{
Create a survival process matrix.
}
\examples{
if (interactive()) {
  bbs_matrix_survival(c(0.5, 0.83, 0.84)) \%*\% rep(100, 3)
}

}
