% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-matrix.R
\name{bbs_matrix_survival_period}
\alias{bbs_matrix_survival_period}
\title{Create survival matrix for each year and period.}
\usage{
bbs_matrix_survival_period(survival)
}
\arguments{
\item{survival}{An array of the survival rates with dimensions period, year and stage. Period represents any subdivision of a year (i.e., week, month, season).}
}
\value{
An array of survival process matrices with dimensions stage, stage, year, period.
}
\description{
Create survival matrix for each year and period.
}
\examples{
survival_rates <- bbs_survival(logit(c(0.94, 0.98)), nyear = 2, nperiod_within_year = 1)
bbs_matrix_survival_period(survival_rates$eSurvival)
}
