% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-population.R
\name{bbs_plot_population}
\alias{bbs_plot_population}
\alias{bbs_plot_population.data.frame}
\alias{bbs_plot_population.bbou_population}
\alias{bbs_plot_population.bbou_simulation}
\alias{bbs_plot_population.bbou_population_caribou}
\title{Plot Population}
\usage{
bbs_plot_population(x, ...)

\method{bbs_plot_population}{data.frame}(x, annual = TRUE, ...)

\method{bbs_plot_population}{bbou_population}(x, annual = TRUE, nperiod_within_year, ...)

\method{bbs_plot_population}{bbou_simulation}(x, annual = TRUE, ...)

\method{bbs_plot_population}{bbou_population_caribou}(x, annual = TRUE, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}

\item{annual}{A flag indicating whether to show annual population (as opposed to monthly).}

\item{nperiod_within_year}{A whole number of the number of periods in a year.}
}
\value{
A ggplot object.

A ggplot object.

A ggplot object.

A ggplot object.
}
\description{
Plots population abundance by period and stage.
}
\section{Methods (by class)}{
\itemize{
\item \code{bbs_plot_population(data.frame)}: Plot population abundance by period and stage for a data frame (abundance data.frame in output \code{\link[=bbs_simulate_caribou]{bbs_simulate_caribou()}}).

\item \code{bbs_plot_population(bbou_population)}: Plot population abundance by period and stage for a matrix (output of \code{\link[=bbs_population_caribou]{bbs_population_caribou()}}).

\item \code{bbs_plot_population(bbou_simulation)}: Plot population abundance by period and stage for a matrix (output of \code{\link[=bbs_simulate_caribou]{bbs_simulate_caribou()}}).

\item \code{bbs_plot_population(bbou_population_caribou)}: Plot population abundance by period and stage for a matrix (output of \code{\link[=bbs_simulate_caribou]{bbs_simulate_caribou()}}).

}}
