% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{bbs_population_caribou}
\alias{bbs_population_caribou}
\title{Simulate Boreal Caribou population from survival, ageing and birth.}
\usage{
bbs_population_caribou(
  survival,
  fecundity,
  adult_females = 1000,
  proportion_adult_female = 0.65,
  proportion_yearling_female = 0.5,
  stochastic = TRUE
)
}
\arguments{
\item{survival}{A list of the Caribou survival rates (output of \code{\link[=bbs_survival_caribou]{bbs_survival_caribou()}})}

\item{fecundity}{A list of the Caribou fecundity rates (output of \code{\link[=bbs_fecundity_caribou]{bbs_fecundity_caribou()}})}

\item{adult_females}{A number of the initial number of adult females in the population.}

\item{proportion_adult_female}{A number between 0 and 1 of the proportion of adults that are female.}

\item{proportion_yearling_female}{A number between 0 and 1 of the proportion of yearlings that are female.}

\item{stochastic}{A flag indicating whether to include demographic stochasticity.}
}
\value{
A matrix of the population by stage and period.
}
\description{
Simulate population projection for Boreal Caribou from key survival and fecundity rates.
}
\details{
This model assumes that survival occurs at the end of each period and survival, ageing and birth occur at the end of each year, in that order.
Initial population is determined by calculating the stable age distribution (output of \code{\link[=bbs_stable_stage_distribution]{bbs_stable_stage_distribution()}}).
Survival and fecundity arguments accept outputs of \code{\link[=bbs_survival_caribou]{bbs_survival_caribou()}} and \code{\link[=bbs_fecundity_caribou]{bbs_fecundity_caribou()}}.
These are converted into processes matrices using \code{\link[=bbs_matrix_survival_period]{bbs_matrix_survival_period()}} and \code{\link[=bbs_matrix_birth_year]{bbs_matrix_birth_year()}} prior to projection.
\code{\link[=bbs_population]{bbs_population()}} is called internally to project population.
Yearling female survival is assumed to be the same as adult female survival.
}
