% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{bbs_population_groups}
\alias{bbs_population_groups}
\title{Assign population into groups}
\usage{
bbs_population_groups(
  population,
  group_size_lambda = 5,
  group_size_theta = 2,
  group_max_proportion = 1/4,
  group_min_size = 2
)
}
\arguments{
\item{population}{A matrix of the population by stage and period (output of \code{\link[=bbs_population]{bbs_population()}} or \code{\link[=bbs_population_caribou]{bbs_population_caribou()}}).}

\item{group_size_lambda}{A number of the lambda value of the gamma-poisson distribution to draw group sizes from.}

\item{group_size_theta}{A number of the theta value of the gamma-poisson distribution to draw group sizes from.}

\item{group_max_proportion}{A number between 0 and 1 of the maximum group size as proportion of the total population.}

\item{group_min_size}{A whole positive number of the minimum group size.}
}
\value{
A list of groups in each period.
}
\description{
In each period the population is assigned into groups.
Group sizes are drawn randomly from a gamma-poisson distribution with specified lambda and theta.
Group sizes are drawn until the cumulative size exceeds the total individuals.
The remaining individuals comprise the final group.
If the remaining number of individuals is < min_size specified, these will be added to the previous group.
Groups are filled by drawing individuals randomly.
}
