% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{bbs_population_groups_pairs}
\alias{bbs_population_groups_pairs}
\title{Assign population into groups by pairs}
\usage{
bbs_population_groups_pairs(
  population,
  group_size_lambda = 5,
  group_size_theta = 2,
  group_max_proportion = 1/4,
  group_min_size = 2,
  recruit_stages = c(1, 2),
  reproductive_female_stages = c(3, 5)
)
}
\arguments{
\item{population}{A matrix of the population by stage and period (output of \code{\link[=bbs_population]{bbs_population()}} or \code{\link[=bbs_population_caribou]{bbs_population_caribou()}}).}

\item{group_size_lambda}{A number of the lambda value of the gamma-poisson distribution to draw group sizes from.}

\item{group_size_theta}{A number of the theta value of the gamma-poisson distribution to draw group sizes from.}

\item{group_max_proportion}{A number between 0 and 1 of the maximum group size as proportion of the total population.}

\item{group_min_size}{A whole positive number of the minimum group size.}

\item{recruit_stages}{A vector of whole numbers indicating the recruit stages (e.g., calf).}

\item{reproductive_female_stages}{A vector of whole numbers indicating the reproductive female stages (e.g., cows).}
}
\value{
A list of groups for each period.
}
\description{
In each period the population is assigned into groups.
Calf-cow pairs are created and assigned to groups together.
Group sizes are drawn randomly from a gamma-poisson distribution with specified lambda and theta.
Group sizes are drawn until the cumulative size exceeds the total individuals. The remaining individuals comprise the final group.
If the remaining number of individuals is < min_size specified, these will be added to the previous group.
Groups are filled by drawing pairs or individuals from the pool.
If there is only space for one individual left in a group, the pool of individuals (i.e., not pairs) will be sampled.
If there are only pairs left in the pool and one individual is required to fill a group, a pair will be broken up.
}
\examples{
survival <- bbs_survival_caribou(0.84)
fecundity <- bbs_fecundity_caribou(0.7)
x <- bbs_population_caribou(survival, fecundity = fecundity, adult_females = 100)
bbs_population_groups_pairs(x)
}
