% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-survey.R
\name{bbs_population_groups_survey}
\alias{bbs_population_groups_survey}
\title{Assign population into groups for each composition survey}
\usage{
bbs_population_groups_survey(
  population,
  month_composition = 9,
  group_size_lambda = 5,
  group_size_theta = 2,
  group_max_proportion = 1/4,
  group_min_size = 2,
  group_coverage = 0.2
)
}
\arguments{
\item{population}{A matrix of the population by stage and period (output of \code{\link[=bbs_population]{bbs_population()}} or \code{\link[=bbs_population_caribou]{bbs_population_caribou()}}).}

\item{month_composition}{A whole number between 1 and 12 of the month that composition surveys take place, relative to the start of the caribou year.}

\item{group_size_lambda}{A number of the lambda value of the gamma-poisson distribution to draw group sizes from.}

\item{group_size_theta}{A number of the theta value of the gamma-poisson distribution to draw group sizes from.}

\item{group_max_proportion}{A number between 0 and 1 of the maximum group size as proportion of the total population.}

\item{group_min_size}{A whole positive number of the minimum group size.}

\item{group_coverage}{A number between 0 and 1 of the proportion of groups sampled.}
}
\value{
A list of observed groups in each composition survey.
}
\description{
See \code{\link[=bbs_population_groups]{bbs_population_groups()}} for details on group assignment.
Groups are sampled in each composition survey in proportion to \code{group_coverage}.
}
