% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{bbs_survival}
\alias{bbs_survival}
\title{Get stochastic survival rates by period, year and stage.}
\usage{
bbs_survival(
  intercept,
  nyear = 10,
  trend = rep(0, length(intercept)),
  annual_sd = rep(0, length(intercept)),
  period_sd = rep(0, length(intercept)),
  annual_period_sd = rep(0, length(intercept)),
  nperiod_within_year = 12
)
}
\value{
An array of survival rates with dimensions period, year, stage.
}
\description{
Each stage can vary by intercept, year trend, annual random effect, period random effect and period within annual random effect.
All values are provided on the log-odds scale.
Variation from random effects is stochastic.
Year is scaled to Year - 1 for trend, which ensures that the intercept is the rate in the first year.
}
\examples{
survival <- bbs_survival(intercept = logit(c(0.94, 0.98, 0.98)), trend = c(0, 0, 0.2))

}
