% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{bbs_survival_caribou}
\alias{bbs_survival_caribou}
\title{Get stochastic Boreal Caribou survival rates by month, year and stage.}
\usage{
bbs_survival_caribou(
  survival_adult_female,
  survival_calf_female = 0.5,
  nyear = 10,
  trend_adult_female = 0,
  annual_sd_adult_female = 0,
  month_sd_adult_female = 0,
  annual_month_sd_adult_female = 0,
  trend_calf_female = 0,
  annual_sd_calf_female = 0,
  month_sd_calf_female = 0,
  annual_month_sd_calf_female = 0,
  yearling_effect = 0
)
}
\arguments{
\item{survival_adult_female}{A number between 0 and 1 of the annual female adult survival.}

\item{survival_calf_female}{A number between 0 and 1 of the annual female calf survival.}

\item{nyear}{A whole number of the number of years.}

\item{trend_adult_female}{A number of the effect of an increase of one year on the log-odds adult female monthly survival.}

\item{annual_sd_adult_female}{A number of the standard deviation of the annual variation in adult female survival on the log-odds monthly survival.}

\item{month_sd_adult_female}{A number of the standard deviation of the monthly variation in adult female survival on the log-odds monthly survival.}

\item{annual_month_sd_adult_female}{A number of the standard deviation of the monthly variation within year variation in adult female survival on the log-odds monthly survival.}

\item{trend_calf_female}{A number of the effect of an increase of one year on the log-odds calf female monthly survival.}

\item{annual_sd_calf_female}{A number of the standard deviation of the annual variation in calf female survival on the log-odds monthly survival.}

\item{month_sd_calf_female}{A number of the standard deviation of the monthly variation in calf female survival on the log-odds monthly survival.}

\item{annual_month_sd_calf_female}{A number of the standard deviation of the monthly variation within year variation in calf female survival on the log-odds monthly survival.}

\item{yearling_effect}{A number of the effect of yearling on the log-odds adult female monthly survival.}
}
\value{
An array of survival rates with dimensions month, year, stage.
}
\description{
Stages include female calves, female yearlings and female adults.
Calf and adult female survival can vary by trend, annual random effect, month random effect and month within annual random effect.
Yearling survival is created as an effect on adult female survival rates.
Variation from random effects is stochastic.
Year is scaled to Year - 1 for trend, which ensures that the intercept is the rate in the first year.
}
\examples{
survival <- bbs_survival_caribou(0.84, survival_calf_female = 0.5)
}
