% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth.R
\name{fecundity_year}
\alias{fecundity_year}
\title{Get stochastic fecundity rates by year and stage.}
\usage{
fecundity_year(intercept, stage, trend, annual_sd, nyear)
}
\arguments{
\item{intercept}{A number of the intercept of the log-odds fecundity.}

\item{stage}{A vector of the effect of stage on the log-odds fecundity. If NA, fecundity is set to 0 for that stage.}

\item{trend}{A number of the effect of an increase of one year on the log-odds fecundity.}

\item{annual_sd}{A number of the standard deviation of the annual variation on the log-odds fecundity.}

\item{nyear}{A whole number of the number of years.}
}
\value{
A matrix of fecundity rates with dimensions year and stage.
}
\description{
Get stochastic fecundity rates by year and stage.
}
\examples{
fecundity <- fecundity_year(4.5, stage = c(0, 0.1, -0.2), 
  trend = 0.1, annual_sd = 0.3, nyear = 5)

}
