% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{\%*b\%}
\alias{\%*b\%}
\title{Stochastic matrix multiplication for age-structured population projection.}
\usage{
x \%*b\% y
}
\arguments{
\item{x}{A matrix with identical number of rows and columns.}

\item{y}{A matrix with one column. If vector, it will be converted to a one-column matrix.}
}
\value{
A matrix.
}
\description{
This assumes that matrix x has identical number of rows and columns; matrix y has one column; and matrix y has the same number of rows as matrix x.
A binomial distribution is used, where size is drawn from the y matrix and prob is drawn from the x matrix.
}
\examples{
if(interactive()){
  population0 <- c(148, 82, 111, 99)
  survival_mat <- matrix_survival(c(0.845, 0.872, 0.859, 0.861))
  set.seed(102)
  survival_mat \%*b\% population0
}

}
