% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth.R
\name{matrix_birth}
\alias{matrix_birth}
\title{Create a birth process matrix.}
\usage{
matrix_birth(
  fecundity,
  female_recruit_state = 1,
  male_recruit_state = 2,
  female_proportion = 0.5
)
}
\arguments{
\item{fecundity}{A vector of the fecundity rates in each state.}

\item{female_recruit_state}{A number indicating the index position of the female recruit state.}

\item{male_recruit_state}{A number indicating the index position of the male recruit state.}

\item{female_proportion}{A number between 0 and 1 indicating the proportion of recruits that are female.}
}
\value{
A matrix of the birth subprocess.
}
\description{
Create a birth process matrix.
}
\examples{
matrix_birth(c(0, 0, 0.2, 0, 0.25, 0)) \%*\% rep(100, 6)
}
