% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth.R
\name{matrix_birth_year}
\alias{matrix_birth_year}
\title{Create a birth process matrix for each year.}
\usage{
matrix_birth_year(
  fecundity,
  female_recruit_state = 1,
  male_recruit_state = 2,
  female_proportion = 0.5
)
}
\arguments{
\item{fecundity}{A matrix of the fecundity rates with dimensions year and state.}
}
\value{
An array with dimensions state, state, year.
}
\description{
Create a birth process matrix for each year.
}
\examples{
#' if (interactive()) {
  birth_year(matrix(c(0, 0.2, 0, 0.3, 0,
                      0, 0.3, 0, 0.35, 0,
                      0, 0.25, 0, 0.3, 0), ncol = 5, byrow = TRUE))
}

}
