% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{matrix_survival}
\alias{matrix_survival}
\title{Create a survival process matrix.}
\usage{
matrix_survival(survival)
}
\arguments{
\item{survival}{A vector of the survival rates in each state.}
}
\value{
A matrix of the survival subprocess.
}
\description{
Create a survival process matrix.
}
\examples{
matrix_survival(c(0.87, 0.84, 0.84, 0.89, 0.9)) \%*\% rep(100, 5)
}
