% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{matrix_survival_period}
\alias{matrix_survival_period}
\title{Create survival matrix for each year and period.}
\usage{
matrix_survival_period(survival)
}
\arguments{
\item{survival}{An array of the survival rates with dimensions period, year and state. Period represents any subdivision of a year (i.e., week, month, season).}
}
\value{
An array of survival process matrices with dimensions state, state, year, period.
}
\description{
Create survival matrix for each year and period.
}
\examples{
#' if (interactive()) {
    survival_rates <- lapply(c(-0.01, 0.01, 0, -0.02, 0.01, 0.01), function(x){
      c(rep(0.98, 4), rep(0.97, 4), rep(0.96, 4)) + x
    })
    survival_rates <- array(unlist(survival_rates), dim = c(4, 3, 6))
    matrix_survival_period(survival_rates)
}

}
