% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for bboutools Functions}
\arguments{
\item{...}{Unused parameters.}

\item{population_init}{A vector of the initial population for each stage.}

\item{birth}{A matrix of the birth process matrix (output of \code{\link[=matrix_birth]{matrix_birth()}}).}

\item{age}{A matrix of the age process matrix (output of \code{\link[=matrix_age]{matrix_age()}}).}

\item{survival}{A matrix of the survival process matrix (output of \code{\link[=matrix_survival]{matrix_survival()}}).}

\item{nyear}{A whole number of the number of years.}

\item{nsims}{A whole number between 1 and 1,000,000 specifying the number of simulated data sets.}

\item{female_recruit_stage}{A positive whole number of the stage representing female recruits.}

\item{male_recruit_stage}{A positive whole number of the stage representing male recruits.}

\item{female_proportion}{A number between 0 and 1 indicating the proportion of recruits that are female.
the number of data sets to simulate.}
}
\description{
Parameter Descriptions for bboutools Functions
}
\keyword{internal}
