% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-survey.R
\name{population_groups_survey}
\alias{population_groups_survey}
\title{Assign population into groups for each composition survey}
\usage{
population_groups_survey(
  population,
  month_composition,
  group_size_year,
  group_size_theta,
  max_group_proportion,
  min_group_size,
  groups_coverage_year
)
}
\arguments{
\item{population}{A population object (output of \code{simulate_population()}).}

\item{group_size_theta}{A number of the theta value of the gamma-poisson distribution to draw groups sizes from.}

\item{max_group_proportion}{A number between 0 and 1 of the maximum group size as proportion of the total population.}

\item{min_group_size}{A whole positive number of the minimum group size.}
}
\value{
A list of observed groups in each composition survey.
}
\description{
See \code{population_groups()} for details on group assignment.
Month composition is month of compsition surveys relative to the start of the caribou year.
Groups are sampled in each composition survey in proportion to groups_coverage_year.
}
