% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_population_period}
\alias{simulate_population_period}
\title{Simulate population with varying rates}
\usage{
simulate_population_period(population_init, birth, age, survival, nsims = 100L)
}
\arguments{
\item{population_init}{A vector of the initial population for each stage.}

\item{birth}{An array of the birth matrices (output of \code{\link[=matrix_birth_year]{matrix_birth_year()}}).}

\item{age}{An age process matrix (output of \link{matrix_age}).}

\item{survival}{An array of the survival matrices (output of \code{\link[=matrix_survival_period]{matrix_survival_period()}}).}

\item{nsims}{A whole number between 1 and 1,000,000 specifying}
}
\value{
A nlist object with projected abundance at each stage, period and simulation.
}
\description{
Simulate population projections given initial population in each stage and survival, ageing, and birth process matrix.
Survival can vary by period (e.g., month) and year, birth can vary by year, and ageing is constant.
This model assumes that survival occurs at the end of each period and survival, ageing and birth occur at the end of each year, in that order.
The dimensions of birth, age and survival process matrices must be identical to the length of the initial population vector.
}
\examples{
if (interactive()) {
  nstage <- 6
  population0 <- rep(100, nstage)
  survival <- matrix_survival(rep(0.87, nstage))
  age <- matrix_age(c(2, 3, 4, 5, 6, 5, 6))
  birth <- matrix_birth(c(0, 0, 0.2, 0, 0.2, 0))
  simulate_population_period(population0, survival = survival, age = age, 
    birth = birth, nyears = 5, nsims = 100)
}
}
