% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-collared.R
\name{survival_collared}
\alias{survival_collared}
\title{Survival of collared females.}
\usage{
survival_collared(
  collared_adult_females,
  month_collar,
  survival_adult_female_month_year,
  probability_uncertain_mortality_month_year,
  probability_uncertain_survival_month_year
)
}
\value{
A data.frame of the number of collared adult females, certain mortalities and uncertain mortalities in each year and month.
}
\description{
The number of collared adult females is 'topped up' at each \code{month_collar}.
Survival rates in each month/year determine the number of mortalities.
The probability of uncertain mortality determines the proportion of mortalities that are uncertain.
The probability of uncertain survival reduces the number of collared adult females without adding mortalities.
}
