% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival_period}
\alias{survival_period}
\title{Get stochastic survival rates by period, year and stage.}
\usage{
survival_period(
  intercept,
  stage,
  trend,
  annual_sd,
  period_sd,
  annual_period_sd,
  nyear,
  nperiod_within_year = 12
)
}
\arguments{
\item{intercept}{A number of the intercept of log-odds periodic survival.}

\item{stage}{A vector of the effect of stage on the log-odds periodic survival.}

\item{trend}{A number of the effect of an increase of one year on the log-odds periodic survival.}

\item{annual_sd}{A number of the standard deviation of the annual variation on the log-odds periodic survival.}

\item{period_sd}{A number of the standard deviation of the periodic (i.e., month, season) variation on the log-odds periodic survival.}

\item{annual_period_sd}{A number of the standard deviation of the periodic variation within year variation on the log-odds periodic survival.}

\item{nyear}{A whole number of the number of years.}

\item{nperiod_within_year}{A whole number of the number of periods within a year.}
}
\value{
An array of survival rates with dimensions period, year, stage.
}
\description{
Get stochastic survival rates by period, year and stage.
}
\examples{
survival <- survival_period(4.5, stage = c(0, 0.1, -0.2), trend = 0.1,
   annual_sd = 0.3, period_sd = 0.2, annual_period_sd = 0.1, nyear = 5, 
   nperiod_within_year = 12)

}
