% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{Template for bbousims Functions}
\arguments{
\item{nyear}{A positive count number of the number of years.}

\item{month_calving}{A count between 1 and 12 of the month that females calf.}

\item{female_fawns}{A count of the number of female fawns at the start of month 1 in year 1.}

\item{male_fawns}{A count of the number of male fawns at the start of month 1 in year 1.}

\item{female_yearlings}{A count of the number of female yearlings at the start of month 1 in year 1.}

\item{male_yearlings}{A count of the number of male yearlings at the start of month 1 in year 1.}

\item{female_adults}{A count of the number of female adults at the start of month 1 in year 1.}

\item{male_adults}{A count of the number of male adults at the start of month 1 in year 1.}

\item{survival_female_fawns_month_year}{A matrix of dimensions 12 x nyear of the monthly survival probability for a female fawn.}

\item{survival_male_fawns_month_year}{A matrix of dimensions 12 x nyear of the monthly survival probability for a male fawn.}

\item{survival_female_yearlings_month_year}{A matrix of dimensions 12 x nyear of the monthly survival probability for a female yearling.}

\item{survival_male_yearlings_month_year}{A matrix of dimensions 12 x nyear of the monthly survival probability for a male yearling.}

\item{survival_female_adults_month_year}{A matrix of dimensions 12 x nyear of the monthly survival probability for a female adult.}

\item{survival_male_adults_month_year}{A matrix of dimensions 12 x nyear of the monthly survival probability for a male adult.}

\item{fawns_per_female_yearling_year}{A vector of length nyear of the expected number of fawns per female yearling.}

\item{fawns_per_female_adult_year}{A vector of length nyear of the expected number of fawns per female adult.}

\item{collared_female_yearlings_month_year}{A matrix of dimensions 12 x nyear of the number of collared female yearlings at the start of the month.}

\item{collared_female_adults_month_year}{A matrix of dimensions 12 x nyear of the number of collared female adults at the start of the month.}

\item{probability_uncertain_mortality_month_year}{A matrix of dimensions 12 x nyear of the monthly probability of the fate of a mortality being uncertain.}

\item{probability_uncertain_survivor_month_year}{A matrix of dimensions 12 x nyear of the monthly probability of the fate of a survivor being uncertain.}

\item{month_census_year}{A vector of length nyear of the census month for each year.}

\item{group_size_year}{A vector of length nyear of the expected group size.}

\item{group_phi_year}{A vector of length nyear of the standard deviation of the extra-Poisson variation in the group size.}

\item{census_groups_year}{A vector of length nyear of the number of groups in the census.}

\item{probability_unsexed_adult_female_year}{A vector of length nyear of the probability of an adult female being unsexed.}

\item{probability_unsexed_adult_male_year}{A vector of length nyear of the probability of an adult male being unsexed.}

\item{mixing_adult_male_year}{A vector of length nyear of the extent of mixing of adult males with adult females.}
}
\description{
Template for bbousims Functions
}
\keyword{internal}
