% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{bb_sims}
\alias{bb_sims}
\title{Simulate Boreal Caribou Data from Model Parameters}
\usage{
bb_sims(
  nyear = 10,
  female_adults = 1000,
  survival_adult_females = logit(0.95),
  survival_fawns = logit(0.95),
  fawns_per_adult_female = 0.3,
  collared_females = 30,
  census_proportion = 0.1,
  month_composition = 9,
  survival_trend = 0,
  survival_amplitude = 0,
  survival_peak = month_composition,
  survival_annual_sd = 0,
  survival_month_sd = 0,
  survival_yearling_effect = 0,
  survival_male_effect = 0,
  fawns_per_yearling_female = fawns_per_adult_female,
  collared_yearling_proportion = 0,
  probability_uncertain_mortality = 0,
  probability_uncertain_survivor = 0,
  group_size = 10,
  group_phi = 10,
  probability_unsexed_adult_female = 0,
  probability_unsexed_adult_male = 0,
  mixing_adult_male = 1
)
}
\arguments{
\item{nyear}{The number of caribou years.}

\item{female_adults}{The number of female adults at the start of the first (calving) month in the first year.}

\item{survival_adult_females}{A probability of the annual female adult survival.}

\item{survival_fawns}{A probability of the annual female fawn survival.}

\item{fawns_per_adult_female}{The expected number of fawns per adult female.}

\item{collared_females}{The total number of collared females at the start of each month.}

\item{census_proportion}{The proportion of the population in the annual census.}

\item{month_composition}{A number between 1 and 12 of the caribou month that composition survey occurs.}

\item{survival_trend}{The effect of an increase of one year on the log-odds monthly survival.}

\item{survival_amplitude}{The amplitude of the effect of the annual cycle in survival.}

\item{survival_peak}{The month of the peak survival in the cycle.}

\item{survival_annual_sd}{The standard deviation of the annual variation on the adult female log-odds monthly survival.}

\item{survival_month_sd}{The standard deviation of the monthly variation on the adult female log-odds monthly survival.}

\item{survival_yearling_effect}{The effect of being a yearling on the adult female log-odds annual survival.}

\item{survival_male_effect}{The effect of being an male on the adult female log-odds annual survival.}

\item{fawns_per_yearling_female}{The expected number of fawns per yearling female.}

\item{collared_yearling_proportion}{The proportion of collared females that are yearlings.}

\item{probability_uncertain_mortality}{The probability of a mortality being uncertain.}

\item{probability_uncertain_survivor}{The probability of a survivor being uncertain.}

\item{group_size}{= 10,}

\item{group_phi}{= 10,}

\item{probability_unsexed_adult_female}{= 0,}

\item{probability_unsexed_adult_male}{= 0,}

\item{mixing_adult_male}{= 1}
}
\value{
A list of three tibbles.
The first named survival has columns year, month, starttotal, mortalitiescertain and mortalitiesuncertain.
The second named recruitment has columns year, month, cows, bulls, unknownadults, yearlings and calves.
And the final named abundance has columns year, month, sex, stage, abundance.
}
\description{
Simulate Boreal Caribou Data from Model Parameters
}
\examples{
bb_sims()
}
