% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{bb_sims}
\alias{bb_sims}
\title{Simulate Boreal Caribou Data from Model Parameters}
\usage{
bb_sims(
  nyear = 10,
  month_composition = 9,
  adult_females = 1000,
  proportion_adult_female = 0.65,
  proportion_yearling_female = 0.5,
  survival_adult_female = 0.85,
  survival_calf = 0.5,
  survival_yearling_effect = 0,
  survival_trend = 0,
  survival_annual_sd = 0,
  survival_month_sd = 0,
  survival_annual_month_sd = 0,
  calves_per_adult_female = 0.8,
  calves_per_adult_female_trend = 0,
  calves_per_adult_female_annual_sd = 0,
  probability_unsexed_adult_female = 0,
  probability_unsexed_adult_male = 0,
  collared_adult_females = 30,
  month_collar = 3L,
  probability_uncertain_mortality = 0,
  probability_uncertain_survivor = 0,
  group_size = 5,
  group_coverage = 0.2
)
}
\arguments{
\item{nyear}{The number of caribou years.}

\item{month_composition}{A number between 1 and 12 of the caribou month that composition survey occurs relative to the calving month.}

\item{survival_yearling_effect}{The effect of being a yearling on the adult female log-odds annual survival.}

\item{survival_trend}{The effect of an increase of one year on the log-odds monthly survival.}

\item{survival_annual_sd}{The standard deviation of the annual variation on the adult female log-odds monthly survival.}

\item{survival_month_sd}{The standard deviation of the monthly variation on the adult female log-odds monthly survival.}

\item{survival_annual_month_sd}{The standard deviation of the monthly variation within year variation on the adult female log-odds monthly survival.}

\item{calves_per_adult_female}{The expected number of calves per adult female.}

\item{probability_unsexed_adult_female}{= 0,}

\item{probability_unsexed_adult_male}{= 0,}

\item{collared_adult_females}{The total number of collared females at the start of each month.}

\item{probability_uncertain_mortality}{The probability of a mortality being uncertain.}

\item{probability_uncertain_survivor}{The probability of a survivor being uncertain.}

\item{group_size}{The expected size of a g,}

\item{survival_adult_females}{A probability of the annual female adult survival.}

\item{survival_calves}{A probability of the annual female calf survival.}

\item{groups_coverage}{The proportion of the groups in the composition survey.}

\item{survival_male_effect}{The effect of being an male on the adult female log-odds annual survival.}

\item{proportion_adult_male}{= 1}
}
\value{
A list of three tibbles.
The first named survival has columns year, month, starttotal, mortalitiescertain and mortalitiesuncertain.
The second named recruitment has columns year, month, cows, bulls, unknownadults, yearlings and calves.
And the final named abundance has columns year, month, sex, stage, abundance.
}
\description{
Simulate Boreal Caribou Data from Model Parameters
}
\examples{
bb_sims()
}
