% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-simple.R
\name{bb_sims_base}
\alias{bb_sims_base}
\title{Simulate Boreal Caribou Data from Base Parameter}
\usage{
bb_sims_base(
  month_composition = 9,
  female_calves = NULL,
  male_calves = female_calves,
  female_yearlings = NULL,
  male_yearlings = female_yearlings,
  female_adults = 1000,
  male_adults = NULL,
  survival_rates = NULL,
  fecundity_rates = NULL,
  collared_female_yearlings_month_year = matrix(5, 12, nyear),
  collared_female_adults_month_year = matrix(30, 12, nyear),
  probability_uncertain_mortality_month_year = matrix(0.1, 12, nyear),
  probability_uncertain_survivor_month_year = matrix(0, 12, nyear),
  group_size_year = rep(10, nyear),
  group_phi_year = rep(2, nyear),
  max_group_proportion = 0.25,
  min_group_size = 3,
  groups_coverage_year = 0.2,
  proportion_adult_male_year = rep(0, nyear),
  probability_unsexed_adult_female_year = rep(0.1, nyear),
  probability_unsexed_adult_male_year = rep(0.1, nyear)
)
}
\arguments{
\item{month_composition}{A number between 1 and 12 of the caribou month that composition survey occurs relative to the calving month.}

\item{female_yearlings}{The number of female yearlings at the start of the first (calving) month in the first year.}

\item{male_yearlings}{The number of male yearlings at the start of the first (calving) month in the first year.}

\item{female_adults}{The number of female adults at the start of the first (calving) month in the first year.}

\item{male_adults}{The number of male adults at the start of the first (calving) month in the first year.}

\item{collared_female_yearlings_month_year}{A matrix of dimensions 12 x nyear of the number of collared female yearlings at the start of the month.}

\item{collared_female_adults_month_year}{A matrix of dimensions 12 x nyear of the number of collared female adults at the start of the month.}

\item{probability_uncertain_mortality_month_year}{A matrix of dimensions 12 x nyear of the monthly probability of the fate of a mortality being uncertain.}

\item{probability_uncertain_survivor_month_year}{A matrix of dimensions 12 x nyear of the monthly probability of the fate of a survivor being uncertain.}

\item{group_size_year}{A vector of length nyear of the expected group size.}

\item{group_phi_year}{A vector of length nyear of the standard deviation of the extra-Poisson variation in the group size.}

\item{groups_coverage_year}{A vector of length nyear of the proportion of the groups in the composition survey.}

\item{proportion_adult_male_year}{A vector of length nyear of the proportion of the adult males in mixed groups in the composition survey.}

\item{probability_unsexed_adult_female_year}{A vector of length nyear of the probability of an adult female being unsexed.}

\item{probability_unsexed_adult_male_year}{A vector of length nyear of the probability of an adult male being unsexed.}

\item{nyear}{The number of caribou years.}

\item{survival_calves_year}{A vector of length nyear of the calf survival from the composition survey to calving.}

\item{survival_female_yearlings_month_year}{A matrix of dimensions 12 x nyear of the monthly survival probability for a female yearling.}

\item{survival_male_yearlings_month_year}{A matrix of dimensions 12 x nyear of the monthly survival probability for a male yearling.}

\item{survival_female_adults_month_year}{A matrix of dimensions 12 x nyear of the monthly survival probability for a female adult.}

\item{survival_male_adults_month_year}{A matrix of dimensions 12 x nyear of the monthly survival probability for a male adult.}

\item{calves_per_female_yearling_year}{A vector of length nyear of the expected number of calves per female yearling during the composition survey.}

\item{calves_per_female_adult_year}{A vector of length nyear of the expected number of calves per female adult during the composition survey.}
}
\value{
A list of three tibbles.
The first named survival has columns year, month, starttotal, mortalitiescertain and mortalitiesuncertain.
The second named recruitment has columns year, month, cows, bulls, unknownadults, yearlings and calves.
And the final named abundance has columns year, month, sex, stage, abundance.
}
\description{
The survival for a month is the survival from the start to the end of a month.
Except for the starting values, the abundance for a month is the abundance at the end of the month.
The number of collared females of a life stage is the minimum of the
number specified and the number of individuals available.
Individual are assigned to composition groups based on their relative proportions
until there are no individuals remaining.
Cows, bulls and yearlings are assigned to groups based on their relative proportions.
If a cow is assigned and there is space for one more individual then a calf is a
For each cow (or yearling if have calves) probability of having a calf.
The model includes demographic stochasticity.
}
\examples{
bb_sims_base()
}
