% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{bbs_demographic_summary}
\alias{bbs_demographic_summary}
\title{Demographic summary}
\usage{
bbs_demographic_summary(
  calves_per_adult_female,
  survival_adult_female,
  survival_calf,
  sex_ratio = 0.5,
  survival_yearling = survival_adult_female
)
}
\arguments{
\item{survival_adult_female}{A number between 0 and 1 of the annual adult female survival.}

\item{survival_calf}{number between 0 and 1 of the annual calf survival.}

\item{sex_ratio}{A number between 0 and 1 of the proportion of females at birth and yearlings.}

\item{survival_yearling}{A number between 0 and 1 of the annual yearling survival.}
}
\value{
A named list of the calf-cow ratio, DeCesare recruitment, leslie matrix, lambda estimate and stable-stage distribution.
}
\description{
Generate calf-cow ratio, DeCesare recruitment, lambda, leslie matrix and stable-stage distribution
from demographic parameters.
}
