% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth.R
\name{bbs_fecundity}
\alias{bbs_fecundity}
\title{Get stochastic fecundity rates by year and stage.}
\usage{
bbs_fecundity(
  intercept,
  trend = rep(0, length(intercept)),
  annual_sd = rep(0, length(intercept)),
  nyear = 10
)
}
\arguments{
\item{intercept}{A vector of the intercept of the log-odds fecundity by stage. If NA, fecundity is set to 0 for all years.}

\item{trend}{A vector of the effect of an increase of one year on the log-odds fecundity by stage.}

\item{annual_sd}{A vector of the standard deviation of the annual variation on the log-odds fecundity by stage.}

\item{nyear}{A whole number of the number of years.}
}
\value{
A matrix of fecundity rates with dimensions year and stage.
}
\description{
Year is scaled to Year - 1 for trend, which ensures that the intercept is the rate in the first year.
}
\examples{
if(interactive()){
  bbs_fecundity(c(NA, logit(0.4)), trend = c(NA, 0.1), annual_sd = c(NA, 0.05))
}

}
