% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-population.R
\name{bbs_plot_population}
\alias{bbs_plot_population}
\alias{bbs_plot_population.data.frame}
\alias{bbs_plot_population.bbou_population}
\alias{bbs_plot_population.bbou_simulation}
\alias{bbs_plot_population.bbou_population_caribou}
\title{Plot Population}
\usage{
bbs_plot_population(x, ...)

\method{bbs_plot_population}{data.frame}(x, annual = TRUE, ...)

\method{bbs_plot_population}{bbou_population}(x, annual = TRUE, nperiod_within_year = 12, ...)

\method{bbs_plot_population}{bbou_simulation}(x, annual = TRUE, alpha = 0.5, ...)

\method{bbs_plot_population}{bbou_population_caribou}(x, annual = TRUE, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}

\item{annual}{A flag indicating whether to show annual population (as opposed to monthly).}

\item{nperiod_within_year}{A whole number of the number of periods in a year.}

\item{alpha}{A number between 0 and 1 of the point transparency.}
}
\value{
A ggplot object.

A ggplot object.

A ggplot object.

A ggplot object.
}
\description{
Plots population abundance by period and stage.
}
\section{Methods (by class)}{
\itemize{
\item \code{bbs_plot_population(data.frame)}: Plot population abundance by period and stage for a data frame (abundance data.frame in output \code{\link[=bbs_simulate_caribou]{bbs_simulate_caribou()}}).

\item \code{bbs_plot_population(bbou_population)}: Plot population abundance by period and stage for a matrix (output of \code{\link[=bbs_population_caribou]{bbs_population_caribou()}}).

\item \code{bbs_plot_population(bbou_simulation)}: Plot population abundance by period and stage for a matrix (output of \code{\link[=bbs_simulate_caribou]{bbs_simulate_caribou()}}).

\item \code{bbs_plot_population(bbou_population_caribou)}: Plot population abundance by period and stage for a matrix (output of \code{\link[=bbs_simulate_caribou]{bbs_simulate_caribou()}}).

}}
\examples{
if (interactive()) {
    survival <- bbs_survival_caribou(0.84)
    fecundity <- bbs_fecundity_caribou(0.7)
    x <- bbs_population_caribou(survival, fecundity = fecundity)
    bbs_plot_population(x[[1]]$abundance)
}
if (interactive()) {
    pop0 <- c(100, 200)
    survival <- bbs_survival(intercept = logit(c(0.95, 0.98)))
    fecundity <- bbs_fecundity(intercept = c(NA, logit(0.4)))
    survival_mat <- bbs_matrix_survival_period(survival$eSurvival)
    birth_mat <- bbs_matrix_birth_year(fecundity$eFecundity)
    age_mat <- bbs_matrix_age(c(2, 2))
    x <- bbs_population(pop0, 
                        birth = birth_mat, 
                        age = age_mat, 
                        survival = survival_mat)
    bbs_plot_population(x)
}
if (interactive()) {
    survival <- bbs_survival_caribou(0.84)
    fecundity <- bbs_fecundity_caribou(0.7)
    x <- bbs_simulate_caribou(survival, fecundity = fecundity, nsims = 3, alpha = 0.7)
    bbs_plot_population(x)
}
if (interactive()) {
    survival <- bbs_survival_caribou(0.84)
    fecundity <- bbs_fecundity_caribou(0.7)
    x <- bbs_population_caribou(survival, fecundity = fecundity)
    bbs_plot_population(x)
}
}
