% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-population.R
\name{bbs_plot_population}
\alias{bbs_plot_population}
\title{Plot population}
\usage{
bbs_plot_population(population, annual = TRUE)
}
\arguments{
\item{population}{A matrix of the population by stage and period (output of \code{\link[=bbs_population]{bbs_population()}} or \code{\link[=bbs_population_caribou]{bbs_population_caribou()}}).}

\item{annual}{A flag indicating whether to show annual population (as opposed to monthly).}
}
\value{
A nlist object with projected abundance at each stage, period and simulation.
}
\description{
Plot simulated population (i.e., from output of \code{bb_simulate_population()}).
}
