% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{bbs_population}
\alias{bbs_population}
\title{Simulate population from survival, ageing and birth}
\usage{
bbs_population(population_init, birth, age, survival, stochastic = TRUE)
}
\arguments{
\item{population_init}{A vector of the initial population for each stage.}

\item{birth}{An array of the birth matrices (output of \code{\link[=bbs_matrix_birth_year]{bbs_matrix_birth_year()}}).}

\item{age}{An age process matrix (output of \link{bbs_matrix_age}).}

\item{survival}{An array of the survival matrices (output of \code{\link[=bbs_matrix_survival_period]{bbs_matrix_survival_period()}}).}

\item{stochastic}{A flag indicating whether to include demographic stochasticity.}
}
\value{
A matrix of the population by stage and period.
}
\description{
Simulate population projections given initial population in each stage and survival, ageing, and birth process matrices.
Survival can vary by period (e.g., month) and year, birth can vary by year, and ageing is constant.
This model assumes that survival occurs at the end of each period and survival, ageing and birth occur at the end of each year, in that order.
The dimensions of birth, age and survival process matrices must be identical to the length of the initial population vector.
}
\examples{
if (interactive()) {
    pop0 <- c(100, 200)
    survival <- bbs_survival(intercept = logit(c(0.95, 0.98)))
    fecundity <- bbs_fecundity(intercept = c(NA, logit(0.4)))
    survival_mat <- bbs_matrix_survival_period(survival$eSurvival)
    birth_mat <- bbs_matrix_birth_year(fecundity$eFecundity)
    age_mat <- bbs_matrix_age(c(2, 2))
    x <- bbs_population(pop0, 
                        birth = birth_mat, 
                        age = age_mat, 
                        survival = survival_mat)
}
}
