% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{bbs_simulate_caribou}
\alias{bbs_simulate_caribou}
\title{Simulate Boreal Caribou Data from key demographic and sampling rates}
\usage{
bbs_simulate_caribou(
  adult_females = 1000,
  nyear = 20,
  survival_adult_female = 0.8,
  survival_calf_female = 0.5,
  calves_per_adult_female = 0.5,
  proportion_adult_female = 0.65,
  proportion_yearling_female = 0.5,
  survival_trend_adult_female = 0,
  survival_trend_calf_female = 0,
  survival_annual_sd_adult_female = 0,
  survival_annual_sd_calf_female = 0,
  survival_month_sd_adult_female = 0,
  survival_month_sd_calf_female = 0,
  survival_annual_month_sd_adult_female = 0,
  survival_annual_month_sd_calf_female = 0,
  calves_per_adult_female_trend = 0,
  calves_per_adult_female_annual_sd = 0,
  collared_adult_females = 30,
  month_collar = 3L,
  probability_uncertain_mortality = 0,
  probability_uncertain_survivor = 0,
  group_size = 5,
  group_coverage = 0.2
)
}
\arguments{
\item{nyear}{The number of caribou years.}

\item{calves_per_adult_female}{The expected number of calves per adult female.}

\item{collared_adult_females}{The total number of collared females at the start of each month.}

\item{probability_uncertain_mortality}{The probability of a mortality being uncertain.}

\item{probability_uncertain_survivor}{The probability of a survivor being uncertain.}

\item{group_size}{The expected size of a g,}

\item{survival_adult_females}{A probability of the annual female adult survival.}

\item{survival_calves}{A probability of the annual female calf survival.}

\item{groups_coverage}{The proportion of the groups in the composition survey.}

\item{survival_trend}{The effect of an increase of one year on the log-odds monthly survival.}

\item{survival_annual_sd}{The standard deviation of the annual variation on the adult female log-odds monthly survival.}

\item{survival_month_sd}{The standard deviation of the monthly variation on the adult female log-odds monthly survival.}

\item{survival_annual_month_sd}{The standard deviation of the monthly variation within year variation on the adult female log-odds monthly survival.}

\item{survival_yearling_effect}{The effect of being a yearling on the adult female log-odds annual survival.}

\item{survival_male_effect}{The effect of being an male on the adult female log-odds annual survival.}

\item{probability_unsexed_adult_female}{= 0,}

\item{probability_unsexed_adult_male}{= 0,}

\item{proportion_adult_male}{= 1}
}
\value{
A list of three tibbles.
The first named survival has columns year, month, starttotal, mortalitiescertain and mortalitiesuncertain.
The second named recruitment has columns year, month, cows, bulls, unknownadults, yearlings and calves.
And the final named abundance has columns year, month, sex, stage, abundance.
}
\description{
Simulate Boreal Caribou Data from key demographic and sampling rates
}
\examples{
bb_sims()
}
