% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{matrix_age}
\alias{matrix_age}
\title{Create an age process matrix.}
\usage{
matrix_age(age = c(2, 3, 3))
}
\arguments{
\item{age}{A vector indicating the stage to age into. If ageing does not occur, the stage should reference itself.}
}
\value{
A matrix of the age subprocess.
}
\description{
Create an age process matrix.
}
\examples{
matrix_age(c(2, 3, 3)) \%*\% c(80, 50, 150)
}
