% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for bboutools Functions}
\arguments{
\item{...}{Unused parameters.}

\item{population_init}{A vector of the initial population for each state.}

\item{birth}{A matrix of the birth process matrix (output of \code{\link[=matrix_birth]{matrix_birth()}}).}

\item{age}{A matrix of the age process matrix (output of \code{\link[=matrix_age]{matrix_age()}}).}

\item{survival}{A matrix of the survival process matrix (output of \code{\link[=matrix_survival]{matrix_survival()}}).}

\item{nyears}{A whole number of the number of years to project population.}
}
\description{
Parameter Descriptions for bboutools Functions
}
\keyword{internal}
