% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for bboutools Functions}
\arguments{
\item{...}{Unused parameters.}

\item{x}{The object.}

\item{population_name}{A string of the population name. This does not affect simulation but can be used as a unique identifier.}

\item{survival_adult_female}{A number between 0 and 1 of the annual female adult survival.}

\item{survival_calf_female}{A number between 0 and 1 of the annual female calf survival.}

\item{calves_per_adult_female}{A number of the calves per adult female.}

\item{proportion_female}{A number between 0 and 1 indicating the proportion of recruits that are female.}

\item{nyear}{A whole number of the number of years.}

\item{nperiod_within_year}{A whole number of the number of periods in a year.}

\item{female_recruit_stage}{A positive whole number of the stage representing female recruits.}

\item{male_recruit_stage}{A positive whole number of the stage representing male recruits. Ignored if NULL.}

\item{stochastic}{A flag indicating whether to include demographic stochasticity.}

\item{month_collar}{A whole number between 1 and 12 of the collaring month.}

\item{collared_adult_females}{A whole positive number of the number of collared adult females.
The number of collared adult females is 'topped up' each year at \code{month_collar}.}

\item{probability_uncertain_mortality}{A number between 0 and 1 of the probability of uncertain mortality.}

\item{probability_uncertain_survival}{A number between 0 and 1 of the probability of uncertain survival.}

\item{probability_unsexed_adult_female}{A number between 0 and 1 of the probability that an adult female is unsexed.}

\item{probability_unsexed_adult_male}{A number between 0 and 1 of the probability that an adult male is unsexed.}

\item{adult_females}{A number of the initial number of adult females in the population.}

\item{proportion_adult_female}{A number between 0 and 1 of the proportion of adults that are female.}

\item{proportion_yearling_female}{A number between 0 and 1 of the proportion of yearlings that are female.}

\item{survival_trend_adult_female}{A number of the effect of an increase of one year on the log-odds adult female survival.}

\item{survival_trend_calf_female}{A number of the effect of an increase of one year on the log-odds calf female survival.}

\item{survival_annual_sd_adult_female}{A number of the standard deviation of the annual variation in log-odds adult female survival.}

\item{survival_annual_sd_calf_female}{A number of the standard deviation of the annual variation in log-odds calf female survival.}

\item{survival_month_sd_adult_female}{A number of the standard deviation of the monthly variation in log-odds adult female survival.}

\item{survival_month_sd_calf_female}{A number of the standard deviation of the monthly variation in log-odds calf female survival.}

\item{survival_annual_month_sd_adult_female}{A number of the standard deviation of the month within annual variation in log-odds adult female survival.}

\item{survival_annual_month_sd_calf_female}{A number of the standard deviation of the month within annual variation in log-odds calf female survival.}

\item{calves_per_adult_female_trend}{A number of the effect of an increase of one year on the log-odds calves per adult female.}

\item{calves_per_adult_female_annual_sd}{A number of the standard deviation of the annual variation on the log-odds calves per adult female.}

\item{population}{A matrix of the population by stage and period (output of \code{\link[=bbs_population]{bbs_population()}} or \code{\link[=bbs_population_caribou]{bbs_population_caribou()}}).}

\item{group_size_lambda}{A number of the lambda value of the gamma-poisson distribution to draw group sizes from.}

\item{group_size_theta}{A number of the theta value of the gamma-poisson distribution to draw group sizes from.}

\item{group_max_proportion}{A number between 0 and 1 of the maximum group size as proportion of the total population.}

\item{group_min_size}{A whole positive number of the minimum group size.}

\item{month_composition}{A whole number between 1 and 12 of the month that composition surveys take place, relative to the start of the biological year.}

\item{group_coverage}{A number between 0 and 1 of the proportion of groups sampled.}

\item{recruit_stages}{A vector of whole numbers indicating the recruit stages (e.g., calf).}

\item{reproductive_female_stages}{A vector of whole numbers indicating the reproductive female stages (e.g., cows).}

\item{survival}{A list of the Caribou survival rates (output of \code{\link[=bbs_survival_caribou]{bbs_survival_caribou()}})}

\item{fecundity}{A list of the Caribou fecundity rates (output of \code{\link[=bbs_fecundity_caribou]{bbs_fecundity_caribou()}})}
}
\description{
Parameter Descriptions for bboutools Functions
}
\keyword{internal}
