% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for bboutools Functions}
\arguments{
\item{...}{Unused parameters.}

\item{survival_adult_female}{A number between 0 and 1 of the annual female adult survival.}

\item{survival_calf_female}{A number between 0 and 1 of the annual female calf survival.}

\item{calves_per_adult_female}{A number of the calves per adult female.}

\item{proportion_female}{A number between 0 and 1 indicating the proportion of recruits that are female.}

\item{nyear}{A whole number of the number of years.}

\item{female_recruit_stage}{A positive whole number of the stage representing female recruits.}

\item{male_recruit_stage}{A positive whole number of the stage representing male recruits. Ignored if NULL.}

\item{stochastic}{A flag indicating whether to include demographic stochasticity.}

\item{population}{A matrix of the population by stage and period (output of \code{\link[=bbs_population]{bbs_population()}} or \code{\link[=bbs_population_caribou]{bbs_population_caribou()}}).}

\item{group_size_lambda}{A number of the lambda value of the gamma-poisson distribution to draw group sizes from.}

\item{group_size_theta}{A number of the theta value of the gamma-poisson distribution to draw group sizes from.}

\item{group_max_proportion}{A number between 0 and 1 of the maximum group size as proportion of the total population.}

\item{group_min_size}{A whole positive number of the minimum group size.}

\item{recruit_stages}{A vector of whole numbers indicating the recruit stages (e.g., calf).}

\item{reproductive_female_stages}{A vector of whole numbers indicating the reproductive female stages (e.g., cows).}
}
\description{
Parameter Descriptions for bboutools Functions
}
\keyword{internal}
