% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-survey.R
\name{population_groups_survey}
\alias{population_groups_survey}
\title{Assign population into groups for each composition survey}
\usage{
population_groups_survey(
  population,
  month_composition,
  group_size_lambda_year,
  group_size_theta_year,
  group_max_proportion,
  group_min_size,
  groups_coverage_year
)
}
\arguments{
\item{population}{A population object (output of \code{simulate_population()}).}
}
\value{
A list of observed groups in each composition survey.
}
\description{
See \code{population_groups()} for details on group assignment.
Month composition is month of compsition surveys relative to the start of the caribou year.
Groups are sampled in each composition survey in proportion to groups_coverage_year.
}
