% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_population_bas_period}
\alias{project_population_bas_period}
\title{Project population from BAS model by period}
\usage{
project_population_bas_period(population_init, birth, age, survival)
}
\arguments{
\item{population_init}{A vector of the initial population for each state.}

\item{birth}{An array of the birth matrices (from \code{\link[=birth_year]{birth_year()}}).}

\item{age}{An array of the age matrices (from \code{\link[=age_year]{age_year()}}).}

\item{survival}{An array of the survival matrices (from \code{\link[=survival_year_month]{survival_year_month()}}).}
}
\value{
An array of the projected population with dimensions state, period and year. Returned object does not include initial population.
}
\description{
project population growth by period given initial population and survival, ageing, and birth subprocesses in that order.
Survival rates vary by year, period and state; fecundity rates vary by year and state; ageing process does not vary with time.
}
\examples{
#' if (interactive()) {
  nstate <- 5
  population0<- rep(100, nstate)
  x1 <- matrix(c(rep(0.98, 12), rep(0.97, 12)), nrow = 12)
  x2 <- matrix(c(rep(0.97, 12), rep(0.96, 12)), nrow = 12)
  x3 <- matrix(c(rep(0.95, 12), rep(0.94, 12)), nrow = 12)
  x4 <- matrix(c(rep(0.99, 12), rep(0.99, 12)), nrow = 12)
  x5 <- matrix(c(rep(0.98, 12), rep(0.96, 12)), nrow = 12)
  
  survival_rates <- array(c(x1, x2, x3, x4, x5), dim = c(12, 2, 5))
  survival <- survival_year_month(survival_rates)
  
  birth_rates <- matrix(c(0, 0.2, 0, 0.3, 0,
                          0, 0.3, 0, 0.35, 0), ncol = 5, byrow = TRUE)
  birth <- birth_year(birth_rates)
  
  age <- age_year(rep(0.5, 2), nstate = 5)
  
  bas_population_growth_period(population0, survival = survival, age = age, birth = birth)
}
}
